requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
django==5.0.7
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.15.2
    # via -r requirements.in
psycopg[binary]==3.2.1
    # via -r requirements.in
psycopg-binary==3.2.1
    # via psycopg
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via psycopg
